BaseObject: allow a specified group of UIDs to change properties without authentication.
This is done with the new set_privileges list, which should be created by the child objects.
An use case for this is on the User object: an user should modify its own properties without being
asked for authentication.

The User object has been modified to use set_privileges for UID 0 and the UID of the user the object
is representing.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>