service and user: add support for adding a newly created user to a set of groups.
A new service method, AddGroupsToUser() has been created. It still isn't exported to DBus as
it needs to be properly security-audited.
This method is now used internally when creating a new user via ShowUserCreationUI(): it's now possible
to specify the groups to add during the method call. The resulting user will be in the supplied groups.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>