Do not load automatically the UI bits.
DBus activatable services start with a clean env, thus core variables needed to specify
the display where the toolkit should draw at are not exported.

By delaying the startup of the UI bits (using a proxy object that loads the module on-demand) we can
specify the display to use at the method call time.

To match this change, methods that need to show an UI will require a new argument, "display".

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>